% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept_invite.R
\name{accept_all_invites}
\alias{accept_all_invites}
\title{Accept an invite}
\usage{
accept_all_invites(git_pat = NULL, verbose = TRUE)
}
\arguments{
\item{git_pat}{Whatever credentials are given are where invites will be accepted from.
If none is supplied, then this will attempt to grab from a git pat set in the environment with usethis::create_github_token().}

\item{verbose}{TRUE/FALSE do you want more progress messages?}
}
\value{
a response from GitHub's API
}
\description{
Given an a user name, accept all invites
}
\examples{
\dontrun{

accept_all_invites()
}
}
