% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{authorize}
\alias{authorize}
\title{Authorize R package to access Github API}
\usage{
authorize(client.id = getOption("github.client.id"),
  client.secret = getOption("github.client.secret"))
}
\description{
This is a function to authorize the R package to access
}

