% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_comments.R
\name{create_issue_comment}
\alias{create_issue_comment}
\title{Create Comment in an Issue}
\usage{
create_issue_comment(owner, repo, number, body = NULL)
}
\arguments{
\item{owner}{The owner of the repository. In a repo name, it would be the part before the slash.}

\item{repo}{The name of the repository}

\item{number}{Issue Number (NOT issue ID)}

\item{body}{The contents of the comment.}
}
\description{
Create Comment in an Issue
}

