% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_milestones.R
\name{create_milestone}
\alias{create_milestone}
\title{Create milestone}
\usage{
create_milestone(owner, repo, title = NULL, description = NULL,
  state = NULL, due_on = NULL)
}
\arguments{
\item{owner}{The owner of the repository. In a repo name, it would be the part before the slash.}

\item{repo}{The name of the repository}

\item{title}{Required. The title of the milestone.}

\item{description}{A description of the milestone.}

\item{state}{The state of the milestone. Either open or closed.}

\item{due_on}{A date in YYYY-MM-DD format}
}
\description{
Create milestone
}

