% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{create_repo_project}
\alias{create_repo_project}
\title{Create a Project in a repository}
\usage{
create_repo_project(owner, repo, project_name, project_desc)
}
\description{
Create a Project in a repository
}

