% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_milestones.R
\name{get_milestone}
\alias{get_milestone}
\title{Get Milestone details}
\usage{
get_milestone(owner, repo, milestone_id)
}
\arguments{
\item{owner}{The owner of the repository. In a repo name, it would be the part before the slash.}

\item{repo}{The name of the repository}

\item{milestone_id}{The milestone_id of the milestone}
}
\description{
Get Milestone details
}

