% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repos.R
\name{get_repos}
\alias{get_repos}
\title{Retrieve all repos for an organization or user}
\usage{
get_repos(owner, how_many = "all", git_pat = NULL, verbose = TRUE)
}
\arguments{
\item{owner}{the name of the organization or user to retrieve a list of repos from. Eg. fhdsl}

\item{how_many}{put the number of how many you would like returned. If all, put "all". By default will return all issues.}

\item{git_pat}{A personal access token from GitHub. Only necessary if the
repository being checked is a private repository.}

\item{verbose}{TRUE/FALSE do you want more progress messages?}
}
\value{
A data frame that contains information about the issues from the given repository
}
\description{
Given a username or organization, retrieve all the repos
}
\examples{
 \dontrun{

repos_df <- get_repos("fhdsl")
}
}
