% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_close_issue}
\alias{gh_close_issue}
\title{Close a GitHub repository's issue.}
\usage{
gh_close_issue(repo, issue_number)
}
\arguments{
\item{repo}{Repository name in the form of "username/reponame" or
"orgname/reponame", e.g. "lwjohnst86/githubr".}

\item{issue_number}{The number of the issue.}
}
\value{
Has not output.
}
\description{
Close a GitHub repository's issue.
}
