% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-requests.R
\name{gh_close_pull_request}
\alias{gh_close_pull_request}
\title{Close a currently open GitHub Pull Request.}
\usage{
gh_close_pull_request(repo, number)
}
\arguments{
\item{repo}{Repository name in the form of "username/reponame" or
"orgname/reponame", e.g. "lwjohnst86/githubr".}

\item{number}{The existing pull request number.}
}
\value{
Nothing.
}
\description{
Close a currently open GitHub Pull Request.
}
