% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_create_label}
\alias{gh_create_label}
\title{Create a GitHub label for a repository.}
\usage{
gh_create_label(repo, name, color, description = NULL)
}
\arguments{
\item{repo}{Repository name in the form of "username/reponame" or
"orgname/reponame", e.g. "lwjohnst86/githubr".}

\item{name}{Name to give the new label.}

\item{color}{Color in hexadecimal. See the list of common \href{https://www.color-hex.com/}{colors}.}

\item{description}{Description to give label.}
}
\value{
No output.
}
\description{
Create a GitHub label for a repository.
}
\examples{

\dontrun{
gh_create_label("lwjohnst86/test-githubr", "Interest", "deadad")
}
}
\seealso{
See \link[githubr:githubr-package]{githubr::githubr-package}.
}
