% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_delete_all_labels}
\alias{gh_delete_all_labels}
\title{Delete all labels from a GitHub repository.}
\usage{
gh_delete_all_labels(repo)
}
\arguments{
\item{repo}{Repository name in the form of "username/reponame" or
"orgname/reponame", e.g. "lwjohnst86/githubr".}
}
\value{
Has no output.
}
\description{
Delete all labels from a GitHub repository.
}
\seealso{
\code{\link[=gh_create_label]{gh_create_label()}} \code{\link[=gh_delete_label]{gh_delete_label()}}
}
