% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_delete_label}
\alias{gh_delete_label}
\title{Delete a GitHub label.}
\usage{
gh_delete_label(repo, label_name)
}
\arguments{
\item{repo}{Repository name in the form of "username/reponame" or
"orgname/reponame", e.g. "lwjohnst86/githubr".}

\item{label_name}{Name of the label.}
}
\value{
No output.
}
\description{
Delete a GitHub label.
}
\examples{

\dontrun{
gh_delete_label("lwjohnst86/test-githubr", "Interest")
}
}
\seealso{
See \link{githubr-package}.
}
