% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\name{gh_delete_release}
\alias{gh_delete_release}
\title{Delete a GitHub release.}
\usage{
gh_delete_release(repo)
}
\arguments{
\item{repo}{Repository name in the form of "username/reponame" or
"orgname/reponame", e.g. "lwjohnst86/githubr".}
}
\value{
Nothing.
}
\description{
Delete a GitHub release.
}
