% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_edit_issue}
\alias{gh_edit_issue}
\title{Edit parts of a GitHub repository's issue.}
\usage{
gh_edit_issue(
  repo,
  issue_number,
  title,
  body,
  state,
  milestone,
  labels,
  assignees
)
}
\arguments{
\item{repo}{Repository name in the form of "username/reponame" or
"orgname/reponame", e.g. "lwjohnst86/githubr".}

\item{issue_number}{The number of the issue.}

\item{title}{The title of the issue.}

\item{body}{Add a description to the issue.}

\item{state}{Whether issue is open or closed.}

\item{milestone}{The milestone (as a number) to add to the issue.}

\item{labels}{Add a label or labels to the issue.}

\item{assignees}{Assign a collaborator to deal with the issue.}
}
\value{
Has not output.
}
\description{
Edit parts of a GitHub repository's issue.
}
