% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{gh_edit_repo}
\alias{gh_edit_repo}
\title{Edit GitHub repository details and settings.}
\usage{
gh_edit_repo(
  repo,
  name,
  description,
  private,
  has_issues,
  has_projects,
  has_wiki,
  default_branch
)
}
\arguments{
\item{repo}{Repository name in the form of "username/reponame" or
"orgname/reponame", e.g. "lwjohnst86/githubr".}

\item{name}{... Need to confirm what this is for.}

\item{description}{Repository description.}

\item{private}{Whether the repository should be public or private. Default is
public.}

\item{has_issues}{Whether repository should have issues.}

\item{has_projects}{Whether repository should have projects.}

\item{has_wiki}{Whether repository should have a wiki.}

\item{default_branch}{What the default branch should be.}
}
\value{
Nothing.
}
\description{
Edit GitHub repository details and settings.
}
