% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_list_issues}
\alias{gh_list_issues}
\title{List issues in GitHub repository.}
\usage{
gh_list_issues(repo, tidied = TRUE)
}
\arguments{
\item{repo}{Repository name in the form of "username/reponame" or
"orgname/reponame", e.g. "lwjohnst86/githubr".}

\item{tidied}{Tidy to data.frame from list?}
}
\value{
A tidy data.frame or a \code{gh_response} class as a list.
}
\description{
List issues in GitHub repository.
}
\examples{
\dontrun{
gh_list_issues("lwjohnst86/test-githubr")
}
}
\seealso{
See \link[githubr:githubr-package]{githubr::githubr-package}.
}
