% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_list_labels}
\alias{gh_list_labels}
\title{List labels in a GitHub repository.}
\usage{
gh_list_labels(repo, tidied = TRUE)
}
\arguments{
\item{repo}{Repository name in the form of "username/reponame" or
"orgname/reponame", e.g. "lwjohnst86/githubr".}

\item{tidied}{Tidy to data.frame from list?}
}
\value{
A tidied \link[tibble:tibble-package]{tibble} or a \code{gh_response} class as
a list.
}
\description{
List labels in a GitHub repository.
}
\examples{

\dontrun{
gh_list_labels("lwjohnst86/test-githubr")

# As a tidied data.frame (or tibble)
tidy(gh_list_labels("lwjohnst86/test-githubr"))
}
}
\seealso{
See \link{githubr-package}.
}
