% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\name{gh_list_tags}
\alias{gh_list_tags}
\title{List all tags in a GitHub repository.}
\usage{
gh_list_tags(repo)
}
\arguments{
\item{repo}{Repository name in the form of "username/reponame" or
"orgname/reponame", e.g. "lwjohnst86/githubr".}
}
\value{
A tidied \link[tibble:tibble-package]{tibble} or a \code{gh_response} class as
a list.
}
\description{
List all tags in a GitHub repository.
}
