% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{gh_list_user_orgs}
\alias{gh_list_user_orgs}
\title{List GitHub organizations an user is a member of.}
\usage{
gh_list_user_orgs(tidied = TRUE)
}
\arguments{
\item{tidied}{Tidied up format via \code{\link[=tidy]{tidy()}}.}
}
\value{
A tidied \link[tibble:tibble-package]{tibble} or a \code{gh_response} class as
a list.
}
\description{
List GitHub organizations an user is a member of.
}
\examples{
\dontrun{
# Requires authenticated user.
gh_list_user_orgs()
}
}
