% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_new_issue}
\alias{gh_new_issue}
\title{Create a GitHub issue in a repository.}
\usage{
gh_new_issue(repo, title, body = NULL, labels = NULL,
  assignees = NULL, milestone = NULL)
}
\arguments{
\item{repo}{Repository name in the form of "username/reponame" or
"orgname/reponame", e.g. "lwjohnst86/githubr".}

\item{title}{The title of the issue.}

\item{body}{Add a description to the issue.}

\item{labels}{Add a label or labels to the issue.}

\item{assignees}{Assign a collaborator to deal with the issue.}

\item{milestone}{The milestone (as a number) to add to the issue.}
}
\value{
Has no output.
}
\description{
Create a GitHub issue in a repository.
}
\examples{

\dontrun{
gh_new_issue("lwjohnst86/test-githubr", title = "Example issue", labels = "bug")
}
}
\seealso{
See \link[githubr:githubr-package]{githubr::githubr-package}.
}
