% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-requests.R
\name{gh_new_pull_request}
\alias{gh_new_pull_request}
\title{Create a GitHub pull request.}
\usage{
gh_new_pull_request(repo, title, head, base, body)
}
\arguments{
\item{repo}{Repository name in the form of "username/reponame" or
"orgname/reponame", e.g. "lwjohnst86/githubr".}

\item{title}{Title for the pull request.}

\item{head}{Branch to merge from (the branch you are wanting merged).}

\item{base}{The "master" branch or the branch you want your changes to be
merged into (the branch you want to merge into).}

\item{body}{The description and details of the pull request.}
}
\value{
Nothing.
}
\description{
Create a GitHub pull request.
}
