% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\name{gh_new_release}
\alias{gh_new_release}
\title{Create a release of a Git tag.}
\usage{
gh_new_release(repo, tag_name, name, body, draft, prerelease)
}
\arguments{
\item{repo}{Repository name in the form of "username/reponame" or
"orgname/reponame", e.g. "lwjohnst86/githubr".}

\item{tag_name}{Name of the tag to create the release from.}

\item{name}{Name of the release.}

\item{body}{Description and details as part of the release.}

\item{draft}{Whether the release should be a draft.}

\item{prerelease}{Whether the release should be a prerelease.}
}
\value{
Nothing.
}
\description{
Create a release of a Git tag.
}
