% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{gh_new_repo}
\alias{gh_new_repo}
\title{Create either a personal or organization repository.}
\usage{
gh_new_repo(repo, description = NULL, private = NULL)
}
\arguments{
\item{repo}{New repository location. In the format of
"username/new_repo_name" or "organization_name/new_repo_name".}

\item{description}{Repository description.}

\item{private}{Whether the repository should be public or private. Default is
public.}
}
\value{
Nothing. Creates a repository on GitHub.
}
\description{
Requires a PAT authentication token.
}
\examples{

\dontrun{
gh_new_repo("lwjohnst86/test-repo")
}

}
\seealso{
See \link[githubr:githubr-package]{githubr::githubr-package}.
}
