\name{githubRepo-class}
\docType{class}
\alias{githubRepo}
\alias{githubRepo-class}
\alias{downloadRepo,githubRepo-method}

\title{Class \code{"githubRepo"}}

\description{
  Class holding information about a specific timepoint of a github repository.
}

\section{Objects from the Class}{
  Objects are generated by calling the \code{getRepo} function.
}

\section{Slots}{
  \describe{
    \item{\code{user}:}{Object of class \code{"character"} specifying the user account that the repository is stored in }
    \item{\code{repo}:}{Object of class \code{"character"} specifying the name of the repository }
    \item{\code{commit}:}{Object of class \code{"character"} specifying the timepoint at which to get information on the repository }
    \item{\code{tree}:}{Object of class \code{"list"} with elements \code{$treeFiles} and \code{$treeUrls} referring to the tree structure present in the repository }
    \item{\code{localPath}:}{Object of class \code{"character"} -- only necessary if file tree is downloaded via \code{downloadRepo} -- specifying the local directory in which the files will be downloaded }
  }
}

\section{Methods}{
  \describe{
    \item{downloadRepo}{\code{signature(repository = "githubRepo")}: ... }
	 }
}

\seealso{
  \code{\link{getRepo-methods}}, \code{\link{downloadRepo-methods}}
}

\keyword{classes}
