% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_comments.R
\name{list_issues_comments}
\alias{list_issues_comments}
\title{List Comments in an Issue}
\usage{
list_issues_comments(owner, repo, number, since = NULL)
}
\arguments{
\item{owner}{The owner of the repository. In a repo name, it would be the part before the slash.}

\item{repo}{The name of the repository}

\item{number}{Issue Number (NOT issue ID)}

\item{since}{Only issues updated at or after this time are returned. Format is YYYY-MM-DD}
}
\description{
List Comments in an Issue
}

