% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_milestones.R
\name{list_milestones}
\alias{list_milestones}
\title{List Milestones in a Repository}
\usage{
list_milestones(owner, repo, state = "open", sort = "due_on",
  direction = "asc")
}
\arguments{
\item{owner}{The owner of the repository. In a repo name, it would be the part before the slash.}

\item{repo}{The name of the repository}

\item{state}{The state of the milestone. Either open, closed, or all.}

\item{sort}{What to sort results by. Either due_on or completeness.}

\item{direction}{The direction of the sort. Either asc or desc.}
}
\description{
List Milestones in a Repository
}

