% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{list_repo_issues}
\alias{list_repo_issues}
\title{List Issues in a Repository}
\usage{
list_repo_issues(owner, repo, milestone = "*", state = "open",
  assignee = "*", creator = NULL, mentioned = NULL, labels = NULL,
  sort = "created", direction = "asc", since = NULL)
}
\arguments{
\item{owner}{The owner of the repository. In a repo name, it would be the part before the slash.}

\item{repo}{The name of the repository}

\item{milestone}{If an integer is passed, it should refer to a milestone by its number field.
If the string * is passed, issues with any milestone are accepted. If the string none is passed,
issues without milestones are returned.}

\item{state}{The state of the milestone. Either open, closed, or all.}

\item{assignee}{Can be the name of a user. Pass in none for issues
with no assigned user, and * for issues assigned to any user.}

\item{creator}{The user that created the issue.}

\item{mentioned}{A user that's mentioned in the issue.}

\item{labels}{A list of comma separated label names. Example: bug,ui,@high}

\item{sort}{What to sort results by. Can be either created, updated, comments. Default: created}

\item{direction}{The direction of the sort. Either asc or desc.}

\item{since}{Only issues updated at or after this time are returned. Format is YYYY-MM-DD}
}
\description{
List Issues in a Repository
}

