% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proj_repo}
\alias{proj_repo}
\title{Project Git remote URL.}
\usage{
proj_repo(remote = "origin")
}
\arguments{
\item{remote}{Remote name. Default is "origin".}
}
\value{
Remote name of R Project.
}
\description{
\lifecycle{experimental}

Get the repository name based on remote name of the current working
directory.
}
