\name{setGithubAuth}
\alias{setGithubAuth}

\title{ DEPRECATED: function used to retrieve github authentication for a specific user }

\description{
  Github has changed their authentication process so this way of setting auth is not longer supported. Read the github blog post explaining why: https://developer.github.com/changes/2014-12-08-removing-authorizations-token/. Please see \code{\link{setGithubToken}} for the updated authorization process for the \code{githubr}.
}

\seealso{
  \code{\link{setGithubToken}}
}
