% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{template_github_request}
\alias{template_github_request}
\title{Basic template request to send to GitHub.}
\usage{
template_github_request(repo, .request, .method = "GET", params_list = NULL)
}
\arguments{
\item{repo}{Repository name in the form of "username/reponame" or
"orgname/reponame", e.g. "lwjohnst86/githubr".}

\item{.request}{Request to GitHub.}

\item{.method}{Method to send, e.g. GET or POST.}

\item{...}{Other parameters to give to the API, e.g.}
}
\value{
Sends a request to GitHub.
}
\description{
Is mostly a thin wrapper around \code{\link[gh:gh]{gh::gh()}}, with some additional conveniences
to simplify GitHub request calls.
}
\seealso{
See \link{githubr-package}.
}
\keyword{internal}
