% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy}
\alias{tidy}
\title{Convert a GitHub response object to a tidied data.frame.}
\usage{
tidy(x, ...)
}
\arguments{
\item{x}{The input to tidy.}

\item{...}{Arguments passed to other methods or cases.}
}
\value{
Returns a data.frame, or if available as a \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Convert a GitHub response object to a tidied data.frame.
}
\examples{

\dontrun{
tidy(gh_list_labels("lwjohnst86/test-githubr"))
}
}
