\name{githubRepo-class}
\docType{class}
\alias{githubRepo}
\alias{githubRepo-class}
\alias{downloadRepo,githubRepo-method}

\title{Class \code{"githubRepo"}}

\description{
  Class holding information about a specific timepoint of a github repository.
}

\section{Objects from the Class}{
  Objects are generated by calling the \code{getRepo} function.
}

\section{Slots}{
  \describe{
    \item{\code{user}:}{Object of class \code{"character"} specifying the user account that the repository is stored in }
    \item{\code{repo}:}{Object of class \code{"character"} specifying the name of the repository }
    \item{\code{ref}:}{Object of class \code{"character"} along with \code{refName}: the type of reference that the repository is trying to capture: one of "branch", "tag", or "commit" }
    \item{\code{refName}:}{Object of class \code{"character"} along with \code{ref}: the name of the branch, tag, or the sha of the requested commit }
    \item{\code{commit}:}{Object of class \code{"character"} specifying the timepoint at which information on the repository was obtained }
    \item{\code{tree}:}{Object of class \code{"data.frame"} referring to the tree structure present in the repository with columns \code{path}, \code{sha} }
  }
}

\section{Methods}{
  \describe{
    \item{\code{\link{sourceRepoFile}}}{\code{signature(repository = "githubRepo", repositoryPath = "character", ...)}}
    \item{\code{\link{view}}}{\code{signature(repository = "githubRepo", repositoryPath = "character")}}
	 }
}

\seealso{
  \code{\link{getRepo}}, \code{\link{sourceRepoFile}}, \code{\link{view}}
}

\keyword{classes}
