% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proj_repo}
\alias{proj_repo}
\title{Get the repository name based on remote name of the current working directory.}
\usage{
proj_repo(remote = "origin")
}
\arguments{
\item{remote}{Remote name. Default is "origin".}
}
\value{

}
\description{
Get the repository name based on remote name of the current working directory.
}
\examples{
proj_repo()
}
