% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy}
\alias{tidy}
\alias{tidy.gh_response}
\title{Convert a \code{gh_response} object to a tidied data.frame.}
\usage{
tidy(x, ...)

\method{tidy}{gh_response}(x, ...)
}
\arguments{
\item{x}{The input to tidy.}

\item{...}{Arguments passed to other methods or cases.}
}
\value{
a \link[tibble:tibble-package]{tibble}.
}
\description{
The class \code{gh_response} is returned from \link[gh:gh]{gh::gh}. The tidy function converts
it from a list to a \link[tibble:tibble-package]{tibble}.
}
\examples{

\dontrun{
# For labels
Labels <- gh_list_labels("lwjohnst86/test-githubr", tidied = FALSE)
Labels
tidy(Labels)
gh_list_labels("lwjohnst86/test-githubr")

Issues <- gh_list_issues("lwjohnst86/test-githubr", tidied = FALSE)
Issues
tidy(Issues)
gh_list_issues("lwjohnst86/test-githubr")
}
}
